+++ translator/src/main/resources/lib/soda/lib/Rec.scala
@@ -9,7 +9,7 @@ package soda.lib
  */
 case class Rec () {
 
-  def foldLeftWhile [A, B, C <: B]  (s: Seq [A], initial_value: C, next_value: (B, A ) => C, condition: (B, A ) => Boolean ): C = {
+  def foldLeftWhile [A, B, C <: (B )]  (s: Seq [A], initial_value: C, next_value: (B, A ) => C, condition: (B, A ) => Boolean ): C = {
 
     lazy val result = rec (s, initial_value )
 
@@ -27,7 +27,7 @@ case class Rec () {
   }
 
 
-  def foldLeft [A, B, C <: B]  (seq: Seq [A], initial_value: C, next_value: (B, A ) => C ): C = {
+  def foldLeft [A, B, C <: (B )]  (seq: Seq [A], initial_value: C, next_value: (B, A ) => C ): C = {
 
     lazy val result = rec (seq, initial_value )