+++ translator/src/test/scala/soda/translator/language/UpperAndLowerBoundDeclarationSpec.scala
@@ -4,9 +4,9 @@ package soda.translator.language
 case class UpperAndLowerBoundDeclarationSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should translate a single upper bound") {
-    lazy val original = "  * BlackBox() extends AbstractBlackBox[A subof AbstractInput]\n"
+    lazy val original = "  * BlackBox() extends AbstractBlackBox[A subtype (AbstractInput)]\n"
 
-    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A <: AbstractInput]\n"
+    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A <: (AbstractInput )]\n"
 
     lazy val obtained = MicroTranslator () .translate_program (original )
     assert (obtained == expected )
@@ -14,18 +14,18 @@ case class UpperAndLowerBoundDeclarationSpec () extends org.scalatest.funsuite.A
 
 
   test ("should translate multiple upper bounds") {
-    lazy val original = "  * BlackBox() extends AbstractBlackBox[A subof AbstractInput] with AbstractDevice[B subof AbstractDeviceInput]\n"
+    lazy val original = "  * BlackBox() extends AbstractBlackBox[A subtype (AbstractInput)] with AbstractDevice[B subtype (AbstractDeviceInput)]\n"
 
-    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A <: AbstractInput] with AbstractDevice [B <: AbstractDeviceInput]\n"
+    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A <: (AbstractInput )] with AbstractDevice [B <: (AbstractDeviceInput )]\n"
 
     lazy val obtained = MicroTranslator () .translate_program (original )
     assert (obtained == expected )
   }
 
   test ("should translate a single lower bound") {
-    lazy val original = "  * BlackBox() extends AbstractBlackBox[A superof AbstractInput]\n"
+    lazy val original = "  * BlackBox() extends AbstractBlackBox[A supertype (AbstractInput)]\n"
 
-    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A >: AbstractInput]\n"
+    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A >: (AbstractInput )]\n"
 
     lazy val obtained = MicroTranslator () .translate_program (original )
     assert (obtained == expected )
@@ -33,9 +33,9 @@ case class UpperAndLowerBoundDeclarationSpec () extends org.scalatest.funsuite.A
 
 
   test ("should translate multiple lower bounds") {
-    lazy val original = "  * BlackBox() extends AbstractBlackBox[A superof AbstractInput] with AbstractDevice[B superof AbstractDeviceInput]\n"
+    lazy val original = "  * BlackBox() extends AbstractBlackBox[A supertype (AbstractInput)] with AbstractDevice[B supertype (AbstractDeviceInput)]\n"
 
-    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A >: AbstractInput] with AbstractDevice [B >: AbstractDeviceInput]\n"
+    lazy val expected = "  case class BlackBox () extends AbstractBlackBox [A >: (AbstractInput )] with AbstractDevice [B >: (AbstractDeviceInput )]\n"
 
     lazy val obtained = MicroTranslator () .translate_program (original )
     assert (obtained == expected )