+++ docs/RELEASE-NOTES.md
@@ -8,8 +8,8 @@
     - replaces `new` command by `@new` annotation, which is only required to translations to Scala 2
     - expands Soda library in multiple files
     - accepts `(` (opening parenthesis) at the end of a line to join lines, like `,` (comma) does
-    - replaces `extends` to declare upper bounds of type parameters by `subof`
-    - includes `superof` to declare lower bounds of type parameters
+    - replaces `extends` to declare upper bounds of type parameters by `subtype`
+    - includes `supertype` to declare lower bounds of type parameters
     - replaces all `package.scala` files by `Package` abstract classes to document the package
   - build: sbt '++ 3.0.0' clean compile test package assembly
   - release: target/scala-3.0.0/soda-0.8.0.jar